make data persistence optional

+++ src/main.rs
@@ -2,10 +2,10 @@
 #![allow(unused_variables)]
 #![allow(dead_code)]
 
+use core::panic;
 use deadpool_postgres::{Manager, ManagerConfig, Pool, RecyclingMethod};
 use duration_string::DurationString;
 use log::{debug, error, info, warn};
-use core::panic;
 use std::borrow::Cow;
 use std::collections::{HashMap, HashSet, VecDeque};
 use std::fs::File;
@@ -35,12 +35,14 @@ include_sql!("src/scripts/library.sql");
 struct Cli {
     #[arg(short, long)]
     streamers: Vec<String>,
+    #[arg(short, long)]
+    persistent: Option<bool>,
 }
 
 async fn watch_task(
     streamer: String,
     mut rx: Receiver<()>,
-    pool: Pool,
+    pool: Option<Pool>,
 ) -> Result<(), Box<dyn error::Error>> {
     info!("Initializing the task for {streamer}");
 
@@ -64,8 +66,10 @@ async fn watch_task(
                                 let channel_id: i32 = msg.channel_id.parse()?;
                                 let user_id: i32 = user_id.parse()?;
 
-                                let db_client = pool.get().await?;
-                                db_client.ban_user(channel_id, user_id, &msg.server_timestamp, &msg.channel_login, &user_login).await?;
+                                if let Some(pool) = pool.as_ref() {
+                                    let db_client = pool.get().await?;
+                                    db_client.ban_user(channel_id, user_id, &msg.server_timestamp, &msg.channel_login, &user_login).await?;
+                                }
 
                                 warn!("{user_login} banned in channel({})", msg.channel_login);
                             }
@@ -73,8 +77,10 @@ async fn watch_task(
                                 let channel_id: i32 = msg.channel_id.parse()?;
                                 let user_id: i32 = user_id.parse()?;
 
-                                let db_client = pool.get().await?;
-                                db_client.timeout_user(channel_id, user_id, &msg.server_timestamp, &msg.channel_login, &user_login, timeout_length.as_secs() as i64).await?;
+                                if let Some(pool) = pool.as_ref() {
+                                    let db_client = pool.get().await?;
+                                    db_client.timeout_user(channel_id, user_id, &msg.server_timestamp, &msg.channel_login, &user_login, timeout_length.as_secs() as i64).await?;
+                                }
 
                                 let duration = DurationString::from(timeout_length);
                                 warn!("{user_login} timeouted in channel({}) for {duration}", msg.channel_login);
@@ -85,8 +91,10 @@ async fn watch_task(
                     ServerMessage::ClearMsg(msg) => {
                         let msg_id = Uuid::parse_str(&msg.message_id)?;
 
-                        let db_client = pool.get().await?;
-                        db_client.delete_message(&msg_id).await?;
+                        if let Some(pool) = pool.as_ref() {
+                            let db_client = pool.get().await?;
+                            db_client.delete_message(&msg_id).await?;
+                        }
 
                         warn!("User's({}) message ({}) deleted in channel({})", msg.sender_login, msg.message_text, msg.channel_login);
                     }
@@ -96,8 +104,10 @@ async fn watch_task(
                         let sender_id: i32 = sender.id.parse()?;
                         let msg_id = Uuid::parse_str(&msg.message_id)?;
 
-                        let db_client = pool.get().await?;
-                        db_client.add_message(channel_id, &msg.channel_login, sender_id, &sender.login, &sender.name, &msg_id, &msg.message_text, &msg.server_timestamp).await?;
+                        if let Some(pool) = pool.as_ref() {
+                            let db_client = pool.get().await?;
+                            db_client.add_message(channel_id, &msg.channel_login, sender_id, &sender.login, &sender.name, &msg_id, &msg.message_text, &msg.server_timestamp).await?;
+                        }
 
                         info!("({}): {}: '{}'", msg.channel_login, sender.name, msg.message_text);
                     }
@@ -118,18 +128,23 @@ async fn main() -> Result<(), Box<dyn error::Error>> {
     }
     info!("streamers: {:?}", cli.streamers);
 
-    let pg_config = Config::new()
-        .host("localhost")
-        .port(5432)
-        .user("twirc")
-        .password("twirc")
-        .dbname("twirc")
-        .to_owned();
-    let mgr_config = ManagerConfig {
-        recycling_method: RecyclingMethod::Fast,
+    let pool = if let Some(true) = cli.persistent {
+        let pg_config = Config::new()
+            .host("localhost")
+            .port(5432)
+            .user("twirc")
+            .password("twirc")
+            .dbname("twirc")
+            .to_owned();
+        let mgr_config = ManagerConfig {
+            recycling_method: RecyclingMethod::Fast,
+        };
+        let mgr = Manager::from_config(pg_config, NoTls, mgr_config);
+        let pool = Pool::builder(mgr).build()?;
+        Some(pool)
+    } else {
+        None
     };
-    let mgr = Manager::from_config(pg_config, NoTls, mgr_config);
-    let pool = Pool::builder(mgr).build()?;
 
     let mut handles = Vec::with_capacity(cli.streamers.len());
     let (tx, rx) = watch::channel(());
